requirements-ai.txt
# AI service specific requirements
# These are for the AI worker containers

# Core AI dependencies
anthropic==0.57.1
openai==1.3.5  # For potential GPT integration
langchain==0.0.340
langchain-anthropic==0.0.1

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Deep Learning (optional, for custom models)
torch==2.1.1 --index-url https://download.pytorch.org/whl/cpu
transformers==4.35.2

# Time series forecasting
statsmodels==0.14.0
prophet==1.1.5
pmdarima==2.0.4

# Vector databases
pinecone-client==2.2.4
chromadb==0.4.18
weaviate-client==3.25.3

# Task queue
celery==5.3.4
celery[redis]==5.3.4
flower==2.0.1

# Data processing
pyarrow==14.0.1
fastparquet==2023.10.1

# Caching
redis==5.0.1
hiredis==2.2.3

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0

# Utilities
tenacity==8.2.3  # Retry logic
cachetools==5.3.2  # LRU cache
python-json-logger==2.0.7
structlog==23.2.0

# Performance
uvloop==0.19.0  # Fast event loop
orjson==3.9.10  # Fast JSON
msgpack==1.0.7  # Binary serialization